third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node-expected.txt

Tests the Timeline API instrumentation of layout invalidations on a deleted node.


Running: testLocalFrame
paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: test://evaluations/0/timeline-paint-with-layout-invalidations-on-deleted-node.js:17}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "BODY"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testSubframe
